pyproject.toml
[tool.poetry]
authors = ["Nicholas Schaub <nick.schaub@nih.gov>", "Camilo Velez <camilo.velez@axleinfo.com>"]
description = "Python API to configure and run Polus Plugins."
license = "License :: OSI Approved :: MIT License"
maintainers = ["Camilo Velez <camilo.velez@axleinfo.com>"]
name = "polus-tabular"
packages = [{include = "polus", from = "src"}]
readme = "README.md"
repository = "https://github.com/PolusAI/tabular-tools"
version = "0.1.1"

[tool.poetry.dependencies]
python = ">=3.9, <3.12"

click = "^8.1.3"
cwltool = "^3.1.20230513155734"
fsspec = "^2023.6.0"
pydantic = ">=1.10.0"
pygithub = "^1.58.2"
python-on-whales = "^0.68.0"
pyyaml = "^6.0"
tqdm = "^4.65.0"
validators = "^0.22.0"
xmltodict = "^0.13.0"

[tool.poetry.group.dev.dependencies]
python = ">=3.9, <3.12"

black = "^23.3.0"
bump2version = "^1.0.1"
datamodel-code-generator = "^0.23.0"
flake8 = "^6.0.0"
fsspec = "^2023.1.0"
mypy = "^1.4.0"
nox = "^2022.11.21"
poetry = "^1.3.2"
pre-commit = "^3.3.3"
pydantic = ">=1.10"
pytest = "^7.3.2"
pytest-benchmark = "^4.0.0"
pytest-cov = "^4.1.0"
pytest-sugar = "^0.9.7"
pytest-xdist = "^3.3.1"
python-on-whales = "^0.68.0"
pyyaml = "^6.0"
ruff = "^0.0.274"
tqdm = "^4.64.1"
xmltodict = "^0.13.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = [
  "--import-mode=importlib",
]
markers = [
  "repo: marks tests that validate plugin.json files in local repo",
]
